+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Frontend
   - Added delete functionality to project budget
   - Added delete functionality to contract invoice processing
+  - Added delete functionality to contract resources
   - Added new dialog modal for forms
   - Removed old for modal